# public/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Connect API
  description: |
    Campus Connect - Social Learning Platform API
    
    Campus Connect is a comprehensive social learning platform for university students featuring:
    - **Study Groups & Squads** - Collaborative learning communities
    - **Learning Sessions** - Scheduled study events and activities
    - **Real-time Chat** - Group messaging and private conversations
    - **Gamification** - Levels, achievements, and leaderboards
    - **Lumina AI** - Intelligent visual learning with animated explanations
    - **Notifications** - Real-time updates for community activity
    
    ## Core Features
    
    ### Social Learning
    - Create and join study groups (squads) with shared interests
    - Organize learning sessions with RSVP tracking
    - Real-time chat rooms for group communication
    - User profiles with topics, interests, and activity tracking
    
    ### Gamification
    - User leveling system (Novice → Learner → Collaborator → Expert → Master)
    - Points earned through engagement (attending events, messaging, creating squads)
    - Achievements and badges
    - Campus-wide leaderboard
    
    ### Lumina AI (Visual Learning)
    Transform natural language queries into educational animations using:
    - **Manim** (3Blue1Brown's animation library)
    - **Gemini 2.5 Flash** (AI script generation)
    - **Daytona.io** (Cloud execution environment)
    - Real-time rendering progress via Server-Sent Events
...
servers:
...
paths:
...
  /users/{id}/level/stats:
    get:
      tags:
        - Users
      summary: Get user level statistics
      description: Get detailed breakdown of user's points and activities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Level statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPoints:
                    type: integer
                    example: 450
                  breakdown:
                    type: object
                    properties:
                      events:
                        type: integer
                        example: 200
                      messages:
                        type: integer
                        example: 150
                      squads:
                        type: integer
                        example: 100
  /users/{id}/activity:
    post:
      tags:
        - Users
      summary: Log user activity
      description: Record user activity for gamification (internal use)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - activityType
              properties:
                activityType:
                  type: string
                  enum:
                    - event_attended
                    - message_sent
                    - squad_created
                    - login
                metadata:
                  type: object
      responses:
        '200':
          description: Activity logged
          content:
            application/json:
              schema:
                type: object
                properties:
                  pointsEarned:
                    type: integer
                  newLevel:
                    type: string
                  leveledUp:
                    type: boolean
...
  /messages:
    get:
      tags:
        - Messages
      summary: Get user's messages
      description: Retrieve direct messages for the current user
      parameters:
        - name: userId
          in: query
          schema:
            type: string
          description: Filter messages by conversation with specific user
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'

    post:
      tags:
        - Messages
      summary: Send direct message
      description: Send a direct message to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipientId
                - content
              properties:
                recipientId:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  # Notifications
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
          example: "dm-123"
        senderId:
          type: string
          example: "alice-123"
        recipientId:
          type: string
          example: "bob-456"
        content:
          type: string
          example: "Hey, are you coming to the study session?"
        read:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication using HTTP-only cookies
security:
  - sessionCookie: []